backend/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
pydantic-settings
uvicorn>=0.29.0,<0.30.0
google-cloud-firestore>=2.16.0,<3.0.0
python-jose
PyJWT
asyncpg
sqlmodel 
sqlalchemy
alembic
pydantic-settings
pwdlib[argon2]
python-dotenv
itsdangerous
psycopg2-binary
authlib
google-cloud-pubsub
google-cloud-run
google-cloud-storage
numpy